moved config types to utils files along with other methods
