+++ examples/README.md
@@ -61,6 +61,22 @@ export { app };
 
 This is the only file that's required to run an example. The code defined in this function is executed each time the example play button is pressed. It takes the example's canvas element from the DOM and usually begins by creating a new PlayCanvas `Application` or `AppBase` using that canvas.
 
+Examples can also contain flags which allow you to define the default configuration for your examples as well as overrides to particular settings such as `deviceType`. Check the possible values to set in `ExampleConfig` in `scripts/utils.mjs` file for the full list.
+
+```js
+// @flag DESCRIPTION This is a description
+// @flag HIDDEN
+// @flag ENGINE
+// @flag INCLUDE_AR_LINK
+// @flag NO_DEVICE_SELECTOR
+// @flag NO_CANVAS
+// @flag NO_MINISTATS
+// @flag WEBGPU_ENABLED
+// @flag WEBGPU_REQUIRED
+import * as pc from 'playcanvas';
+...
+```
+
 You can load external scripts into an example using the `loadES5` function as follows:
 
 ```js
@@ -106,22 +122,6 @@ The data observer used in the `controls` function will be made available as an i
 import { data } from '@examples/observer';
 
 console.log(data.get('flash'));
-```
-
-### `config.mjs`
-
-This file allows you to define the default configuration for your examples as well as overrides to particular settings such as `deviceType`. Check the config options `ExampleConfig` in `types.mjs` file for the full list.
-
-```js
-/**
- * @type {import('../../../../types.mjs').ExampleConfig}
- */
-export default {
-    NO_CANVAS: true,
-    NO_MINISTATS: true,
-    WEBGPU_ENABLED: true,
-    WEBGPU_REQUIRED: true
-};
 ```
 
 ### Additional files
