--- examples/types.mjs
@@ -1,12 +1,0 @@
-/**
- * @typedef {object} ExampleConfig
- * @property {string} [DESCRIPTION] - The example description.
- * @property {boolean} [HIDDEN] - The example is hidden on production.
- * @property {'DEVELOPMENT' | 'PERFORMANCE' | 'DEBUG'} [ENGINE] - The engine type.
- * @property {boolean} [INCLUDE_AR_LINK] - Include AR link png.
- * @property {boolean} [NO_DEVICE_SELECTOR] - No device selector.
- * @property {boolean} [NO_CANVAS] - No canvas element.
- * @property {boolean} [NO_MINISTATS] - No ministats.
- * @property {boolean} [WEBGPU_ENABLED] - If webGPU is enabled.
- * @property {boolean} [WEBGPU_REQUIRED] - If webGPU is required (overrides {@link WEBGPU_ENABLED} ).
- */
